+++ Backend/src/core/types/email.types.ts
@@ -1,3 +1,4 @@
+
 export interface EmailListDTO {
     _id: string;
     subject?: string;